@@ -11,10 +11,10 @@
         }
     ],
     "require": {
-        "php": "^7.0",
-        "symfony/symfony": "~3.0",
+        "php": "^7.4 || ^8.0",
+        "symfony/symfony": "~3.0 || ~4.0",
         "ongr/elasticsearch-bundle": "~5.0",
-        "jms/serializer-bundle": "^2.0"
+        "jms/serializer-bundle": "^2.0 || ^3.0"
     },
     "require-dev": {
         "monolog/monolog": "~1.0",
