@@ -11,9 +11,7 @@
 
 namespace ONGR\FilterManagerBundle\DependencyInjection;
 
-use ONGR\ElasticsearchDSL\Aggregation\TermsAggregation;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
-use Symfony\Component\Config\Definition\Builder\ParentNodeDefinitionInterface;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -25,10 +23,10 @@ class Configuration implements ConfigurationInterface
     /**
      * {@inheritdoc}
      */
-    public function getConfigTreeBuilder()
+    public function getConfigTreeBuilder(): TreeBuilder
     {
-        $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('ongr_filter_manager');
+        $treeBuilder = new TreeBuilder('ongr_filter_manager');
+        $rootNode = $treeBuilder->getRootNode();
 
         $this->addManagersSection($rootNode);
         $this->addFiltersSection($rootNode);
@@ -36,10 +34,8 @@ class Configuration implements ConfigurationInterface
         return $treeBuilder;
     }
 
-    /**
-     * @param ArrayNodeDefinition $rootNode
-     */
-    private function addManagersSection(ArrayNodeDefinition $rootNode)
+
+    private function addManagersSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
             ->children()
@@ -65,10 +61,8 @@ class Configuration implements ConfigurationInterface
             ->end();
     }
 
-    /**
-     * @param ArrayNodeDefinition $rootNode
-     */
-    private function addFiltersSection(ArrayNodeDefinition $rootNode)
+
+    private function addFiltersSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
             ->children()
@@ -98,14 +92,8 @@ class Configuration implements ConfigurationInterface
             ->end();
     }
 
-    /**
-     * Builds relations config tree for given relation name.
-     *
-     * @param string $relationType
-     *
-     * @return ArrayNodeDefinition
-     */
-    private function buildRelationsTree($relationType)
+
+    private function buildRelationsTree(string $relationType): ArrayNodeDefinition
     {
         $filter = new ArrayNodeDefinition($relationType);
 
